v0.16.13
I made a fridge. I told myself I have to make all of the furniture in this room by the end of this month or else I'm the big gay.
Jokes aside, I've been busy with work and school, two large personal projects and I'm sick so I put this game to the side to work on recovering. And maybe catch up in my programming courses.